moved function to task_utils
